+++ app/src/main/java/com/keepingstock/data/entities/ItemWithTags.kt
@@ -1,8 +1,4 @@
-package com.keepingstock.data
-
-import androidx.room.Embedded
-import androidx.room.Junction
-import androidx.room.Relation
+package com.keepingstock.data.entities
 
 /*
 * This code was generated with the help of Android Basics with Compose course.
@@ -10,6 +6,10 @@ import androidx.room.Relation
 * https://developer.android.com/training/data-storage/room/relationships/many-to-many
 * documents the sample code that led to my code.
 */
+
+import androidx.room.Embedded
+import androidx.room.Junction
+import androidx.room.Relation
 
 data class ItemWithTags (
     @Embedded val item: ItemEntity,
